requirements.txt
streamlit
openai
python-dotenv
pypdf
elevenlabs
httpx